adjust sructure
