+++ src/exceptions/AuthenticationTokenMissingException.ts
