+++ src/exceptions/CategoryNotFoundException.ts
