+++ src/exceptions/NoClassException.ts
