+++ src/exceptions/PostNotFoundException.ts
