+++ src/exceptions/UserWithThatEmailAlreadyExistsException.ts
