+++ src/exceptions/WrongAuthenticationTokenException.ts
