+++ src/repositories/teachingPlatform/teachingPlatform.entity.ts
