+++ src/routes.ts
@@ -1,0 +1,6 @@
+import express, { Router } from "express";
+import classRoutes from "./repositories/class/class.route";
+
+const routes = Router();
+
+routes.use("/class", classRoutes);
